requirements.txt
# Core dependencies
streamlit>=1.42.0
transformers>=4.18.0
torch>=1.1.0

# Google AI and Authentication
google-generativeai>=0.3.0
streamlit-google-auth>=0.1.0

# Database and Storage

gspread>=5.0.0
oauth2client>=4.1.3

# HTTP requests and n8n integration
requests>=2.28.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Additional utilities
python-dotenv>=1.0.0
